refactor(acceptance-test): 각 기능 별 인수테스트 분리 및 랜덤 포트 사용
